♻️REFACTOR: TotalDiscount, ChristmasTotalDiscount 패키지 변경.
