+++ src/test/java/christmas/model/domain/event/EventBenefitsTest.java
@@ -4,7 +4,7 @@ import static christmas.model.domain.order.RequestOrdersTest.makeRequestOrdersWi
 import static org.assertj.core.api.Assertions.assertThat;
 
 import christmas.model.domain.date.Date;
-import christmas.model.domain.event.benefit.ChristmasTotalDiscount;
+import christmas.model.domain.event.benefit.totaldiscount.ChristmasTotalDiscount;
 import christmas.model.domain.event.benefit.EventBenefits;
 import christmas.model.domain.event.strategy.ChristmasEventStrategy;
 import christmas.model.domain.order.RequestOrders;
